Refactor generateData to it's own folder. Separate modules
